.history/requirements_20260105133357.txt
asgiref==3.10.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-ckeditor==6.7.3
django-embed-video==1.4.10
django-jazzmin==3.0.1
django-jet==1.0.8
django-js-asset==3.1.2
django-mptt==0.18.0
django-multiselectfield==1.0.1
Faker==37.11.0
google-auth==2.41.1
idna==3.11
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
